Nov 18,2020

+++ Basic/n factorial recursion.cpp
@@ -2,19 +2,18 @@
 using namespace std;
 int n;
 
-int fact(int n)
+long long int fact(long long int n)
 {
     if(n<=1)
         return 1;
-
     else
         return n*fact(n-1);
 }
 int main()
 {
-    int n;
+    long long int n;
     cin>>n;
-    int res = fact(n);
+    long long int res = fact(n);
     cout<<"factorial of "<<n<<" is: "<<res<<endl;
     return 0;
 }
